requirements/base.txt
dj-database-url==2.1.0
Django==5.0.1
phonenumbers==8.13.27
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework==3.14.0
pathlib==1.0.1
celery==5.3.6
flower==2.0.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
Faker==22.1.0
pillow==10.2.0
pyotp==2.9.0
Markdown==3.5.1